+++ examples/simulations/lorenz.py
@@ -21,13 +21,17 @@ line.add_shadow("x",   3, alpha=0.2)
 line.add_shadow("z", -25, alpha=0.2)
 
 pt = Point(pts[0]).color("red4").ps(12)
-pt.add_trail(lw=4).add_shadow("x", 3, alpha=0.5)
-pt.trail.add_shadow("x", 3, alpha=0.5)
+pt.add_trail(lw=4)
+pt.add_shadow("x", 3, alpha=0.5)
+pt.trail.add_shadow("x",  3, alpha=0.5)
+pt.trail.add_shadow("z",-25, alpha=0.5)
 
 def loop_func(event): # move the point
     if len(pts) > 0:
         pos = pts.pop(0)
-        pt.pos(pos).update_trail().update_shadows()
+        pt.pos(pos)
+        pt.update_trail()
+        pt.update_shadows()
         plt.render()
 
 plt = Plotter(axes=dict(xygrid=False))